feat: extract errors routes to routes/errors.py
