+++ templates/base.html
@@ -735,7 +735,7 @@
         </a>
       </li>
       <li class="nav-item">
-        <a href="{{ url_for('view_errors') }}" class="nav-link {% if request.endpoint == 'view_errors' %}active{% endif %}" title="Errors">
+        <a href="{{ url_for('errors.view_errors') }}" class="nav-link {% if request.endpoint == 'errors.view_errors' %}active{% endif %}" title="Errors">
           <i class="bi bi-bug"></i><span class="sb-label">Errors</span>
         </a>
       </li>
