+++ templates/errors_help.html
@@ -10,7 +10,7 @@
     <h4 class="fw-semibold mb-1">Errors Help</h4>
     <p class="text-secondary mb-0">Error tracking, one-click resolve, and grouped deduplication mode.</p>
   </div>
-  <a class="btn btn-outline-secondary btn-sm" href="{{ url_for('view_errors') }}" title="Back to Errors">
+  <a class="btn btn-outline-secondary btn-sm" href="{{ url_for('errors.view_errors') }}" title="Back to Errors">
     <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to Errors</span>
   </a>
 </div>
