+++ templates/metrics_anomaly.html
@@ -53,7 +53,7 @@
        data-ai-risk="low"
        data-ai-confirm="false"
        title="Open Related Errors"
-       href="{{ url_for('view_errors', service=service, from_ts=from_ts, to_ts=to_ts) }}">
+       href="{{ url_for('errors.view_errors', service=service, from_ts=from_ts, to_ts=to_ts) }}">
       <i class="bi bi-bug me-1"></i><span class="metrics-anomaly-btn-label">Open Related Errors</span>
     </a>
     {% endif %}
@@ -233,7 +233,7 @@
               </a>
               {% elif r.related_target == 'errors' %}
               <a class="btn btn-sm btn-outline-info"
-                 href="{{ url_for('view_errors', service=r.service, from_ts=r.time, window_s=60) }}">
+                 href="{{ url_for('errors.view_errors', service=r.service, from_ts=r.time, window_s=60) }}">
                 Errors
               </a>
               {% endif %}
