+++ templates/summary.html
@@ -78,7 +78,7 @@
     </a>
   </div>
   <div class="col-6 col-md-4 col-lg-2">
-     <a href="{{ url_for('view_errors') }}" class="text-decoration-none"
+     <a href="{{ url_for('errors.view_errors') }}" class="text-decoration-none"
        data-ai-action-id="summary.nav.errors"
        data-ai-action-type="navigate"
        data-ai-handler="navigate"
@@ -176,7 +176,7 @@
           <span><i class="bi bi-bug me-1 text-danger"></i>Recent Open Errors</span>
           {% if recent_errors or recent_logs %}<span id="summary-errors-tz-badge-btn" class="sobs-tz-header-badge" role="button" tabindex="0" title="Display timezone - click to change"><i class="bi bi-clock"></i><span id="summary-errors-tz-badge-label">UTC</span></span>{% endif %}
         </div>
-        <a href="{{ url_for('view_errors') }}" class="btn btn-sm btn-outline-secondary" title="View all"><i class="bi bi-arrow-right me-1"></i><span class="summary-btn-label">View all</span></a>
+        <a href="{{ url_for('errors.view_errors') }}" class="btn btn-sm btn-outline-secondary" title="View all"><i class="bi bi-arrow-right me-1"></i><span class="summary-btn-label">View all</span></a>
       </div>
       <div class="card-body p-0">
         {% if recent_errors %}
